refactor: :recycle: assistant -> survey 변경에 따른 폴더 구조 & 명칭 변경
