--- src/pages/assistant/constants/healthInfo/index.ts
@@ -1,88 +1,0 @@
-export const diseaseList = [
-  {
-    value: '고혈압',
-    label: '고혈압',
-  },
-  {
-    value: '고지혈증',
-    label: '고지혈증',
-  },
-  {
-    value: '뇌혈관질환',
-    label: '뇌혈관질환',
-  },
-  {
-    value: '심장질환',
-    label: '심장질환',
-  },
-  {
-    value: '당뇨병',
-    label: '당뇨병',
-  },
-  {
-    value: '갑상선질환',
-    label: '갑상선질환',
-  },
-  {
-    value: '위장관질환',
-    label: '위장관질환',
-  },
-  {
-    value: '파킨슨',
-    label: '파킨슨',
-  },
-  {
-    value: '치매, 인지장애',
-    label: '치매, 인지장애',
-  },
-  {
-    value: '수면장애',
-    label: '수면장애',
-  },
-  {
-    value: '우울증/불안장애',
-    label: '우울증/불안장애',
-  },
-  {
-    value: '신장질환',
-    label: '신장질환',
-  },
-  {
-    value: '간질환',
-    label: '간질환',
-  },
-  {
-    value: '비뇨·생식기질환(전립선비대증, 자궁내막염, 방광염 등)',
-    label: '비뇨·생식기질환(전립선비대증, 자궁내막염, 방광염 등)',
-  },
-  {
-    value: '암질환',
-    label: '암질환',
-  },
-  {
-    value: '뇌경색',
-    label: '뇌경색',
-  },
-];
-
-export const isAllergyTypes = [
-  {
-    value: true,
-    label: '알레르기 있음',
-  },
-  {
-    value: false,
-    label: '알레르기 없음',
-  },
-];
-
-export const isMedicineTypes = [
-  {
-    value: true,
-    label: '약물 부작용 있음',
-  },
-  {
-    value: false,
-    label: '약물 부작용 없음',
-  },
-];
