--- src/pages/assistant/constants/lifeInfo/index.ts
@@ -1,140 +1,0 @@
-export const isSmokingTypes = [
-  {
-    label: '흡연',
-    value: true,
-  },
-  {
-    label: '비흡연',
-    value: false,
-  },
-];
-export const smokingAmount = [
-  {
-    label: '1갑',
-    value: '1갑',
-  },
-  {
-    label: '2갑',
-    value: '2갑',
-  },
-  {
-    label: '3갑 이상',
-    value: '3갑 이상',
-  },
-];
-export const isDrinkingTypes = [
-  {
-    label: '음주',
-    value: true,
-  },
-  {
-    label: '비음주',
-    value: false,
-  },
-];
-export const drkingWeeklyCounts = [
-  {
-    label: '주 1회',
-    value: '주 1회',
-  },
-  {
-    label: '주 2회',
-    value: '주 2회',
-  },
-  {
-    label: '주 3회',
-    value: '주 3회',
-  },
-  {
-    label: '주 4회',
-    value: '주 4회',
-  },
-  {
-    label: '주 5회 이상',
-    value: '주 5회 이상',
-  },
-];
-export const dailyEatingTypes = [
-  {
-    label: '하루 한 끼 규칙적 식사',
-    value: '하루 한 끼 규칙적 식사',
-  },
-  {
-    label: '하루 두끼 규칙적 식사',
-    value: '하루 두끼 규칙적 식사',
-  },
-  {
-    label: '하루 세끼 규칙적 식사',
-    value: '하루 세끼 규칙적 식사',
-  },
-  {
-    label: '불규칙적 식사',
-    value: '불규칙적 식사',
-  },
-];
-export const exerciseWeeklyCounts = [
-  {
-    label: '주 1회',
-    value: '주 1회',
-  },
-  {
-    label: '주 2회',
-    value: '주 2회',
-  },
-  {
-    label: '주 3회',
-    value: '주 3회',
-  },
-  {
-    label: '주 4회',
-    value: '주 4회',
-  },
-  {
-    label: '주 5회 이상',
-    value: '주 5회 이상',
-  },
-  {
-    label: '운동 안 함',
-    value: '운동 안 함',
-  },
-];
-export const isAloneTypes = [
-  {
-    label: '독거',
-    value: true,
-  },
-  {
-    label: '동거',
-    value: false,
-  },
-];
-export const medicinetakingMembers = [
-  {
-    label: '본인',
-    value: '본인',
-  },
-  {
-    label: '배우자',
-    value: '배우자',
-  },
-  {
-    label: '자녀',
-    value: '자녀',
-  },
-  {
-    label: '친인척',
-    value: '친인척',
-  },
-  {
-    label: '친구',
-    value: '친구',
-  },
-  {
-    label: '요양보호사 또는 돌봄종사자',
-    value: '요양보호사 또는 돌봄종사자',
-  },
-  {
-    label: '기타',
-    value: '기타',
-  },
-];
