--- src/pages/assistant/constants/modal/index.ts
@@ -1,5 +1,0 @@
-export type CounselAssistantDialogTypes =
-  | 'REGISTER'
-  | 'TEMP_SAVE'
-  | 'EXIT'
-  | null;
