--- src/pages/assistant/layout/AssistantLayout.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { Outlet } from 'react-router-dom';
-
-const AssistantLayout: React.FC = () => {
-  return (
-    <div>
-      <Outlet />
-    </div>
-  );
-};
-
-export default AssistantLayout;
